RLearn/second/requirements.txt
# please use 'pip install -r requirements.txt' to install packages
libsvm-official
numpy
pandas
matplotlib